+++ src/libromdata/Console/n64_structs.h
@@ -42,6 +42,7 @@ extern "C" {
 typedef union PACKED _N64_RomHeader {
 	struct {
 		union {
+			// [0x000]
 			// NOTE: Technically, the first two DWORDs
 			// are initialization settings, but in practice,
 			// they're usually identical for all N64 ROMs.
@@ -52,12 +53,14 @@ typedef union PACKED _N64_RomHeader {
 			};
 		};
 
-		uint32_t entrypoint;
-		uint32_t release;	// ???
-		uint64_t checksum;
-		uint8_t reserved1[8];
-		char title[0x14];	// Title. (cp932)
-		uint8_t reserved[7];
+		uint32_t entrypoint;	// [0x008]
+		uint8_t release[4];	// [0x00C] OS version. Format: 00 00 AA BB
+					//         AA is decimal; BB is ASCII.
+					//         OoT is 00 00 14 49 == OS 20I
+		uint32_t crc[2];	// [0x010] Two CRCs.
+		uint8_t reserved1[8];	// [0x018]
+		char title[0x14];	// [0x020] Title. (cp932)
+		uint8_t reserved[7];	// [0x034]
 		char id4[4];		// Game ID.
 		uint8_t revision;	// Revision.
 	};