+++ NEWS.md
@@ -35,6 +35,7 @@
   * DMG: Added more entry point formats.
   * DirectDrawSurface: Untested support for Xbox One textures.
   * NintendoDS: Improved DSi flag decoding; added key index decoding.
+  * N64: Fix CRC fields; added Release field.
 
 * New compressed texture formats:
   * Ericsson ETC1 and ETC2